+++ javav2/usecases/creating_amazon_cognito_app/src/main/java/com/example/handlingformsubmission/DynamoDBEnhanced.java
@@ -1,35 +1,25 @@
 /*
    Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
-   This file is licensed under the Apache License, Version 2.0 (the "License").
-   You may not use this file except in compliance with the License. A copy of
-   the License is located at
-    http://aws.amazon.com/apache2.0/
-   This file is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
-   CONDITIONS OF ANY KIND, either express or implied. See the License for the
-   specific language governing permissions and limitations under the License.
+   SPDX-License-Identifier: Apache-2.0
 */
 
-
 package com.example.handlingformsubmission;
-
-import static software.amazon.awssdk.enhanced.dynamodb.mapper.StaticAttributeTags.primaryPartitionKey;
 
 import software.amazon.awssdk.auth.credentials.EnvironmentVariableCredentialsProvider;
 import software.amazon.awssdk.enhanced.dynamodb.DynamoDbEnhancedClient;
 import software.amazon.awssdk.enhanced.dynamodb.DynamoDbTable;
 import software.amazon.awssdk.enhanced.dynamodb.TableSchema;
-import software.amazon.awssdk.enhanced.dynamodb.mapper.StaticTableSchema;
-
 import software.amazon.awssdk.enhanced.dynamodb.model.PutItemEnhancedRequest;
 import software.amazon.awssdk.regions.Region;
 import software.amazon.awssdk.services.dynamodb.DynamoDbClient;
-import software.amazon.awssdk.services.dynamodb.model.ProvisionedThroughput;
 import org.springframework.stereotype.Component;
+import software.amazon.awssdk.services.dynamodb.model.DynamoDbException;
 
 @Component("DynamoDBEnhanced")
 public class DynamoDBEnhanced {
-    // Uses the Enhanced Client to inject a new post into a DynamoDB table
-    public void injectDynamoItem(Greeting item){
+
+     public void injectDynamoItem(Greeting item){
+
         Region region = Region.US_EAST_1;
         DynamoDbClient ddb = DynamoDbClient.builder()
                 .region(region)
@@ -41,7 +31,6 @@ public class DynamoDBEnhanced {
                     .dynamoDbClient(ddb)
                     .build();
 
-            //Create a DynamoDbTable object
             DynamoDbTable<GreetingItems> mappedTable = enhancedClient.table("Greeting", TableSchema.fromBean(GreetingItems.class));
             GreetingItems gi = new GreetingItems();
             gi.setName(item.getName());
@@ -55,55 +44,9 @@ public class DynamoDBEnhanced {
 
             mappedTable.putItem(enReq);
 
-        } catch (Exception e) {
-            e.getStackTrace();
-        }
-
-    }
-
-    public class GreetingItems {
-
-        //Set up Data Members that correspond to columns in the Work table
-        private String id;
-        private String name;
-        private String message;
-        private String title;
-
-        public GreetingItems()
-        {
-        }
-
-        public String getId() {
-            return this.id;
-        }
-
-        public void setId(String id) {
-            this.id = id;
-        }
-
-        public String getName() {
-            return this.name;
-        }
-
-        public void setName(String name) {
-            this.name = name;
-        }
-
-        public String getMessage(){
-            return this.message;
-        }
-
-        public void setMessage(String message){
-            this.message = message;
-        }
-
-
-        public String getTitle() {
-            return this.title;
-        }
-
-        public void setTitle(String title) {
-            this.title = title;
+        } catch (DynamoDbException e) {
+            System.err.println(e.getMessage());
+            System.exit(1);
         }
     }
 }
