+++ javav2/usecases/creating_amazon_cognito_app/src/main/java/com/example/handlingformsubmission/GreetingController.java
@@ -1,14 +1,7 @@
 /*
    Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
-   This file is licensed under the Apache License, Version 2.0 (the "License").
-   You may not use this file except in compliance with the License. A copy of
-   the License is located at
-    http://aws.amazon.com/apache2.0/
-   This file is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
-   CONDITIONS OF ANY KIND, either express or implied. See the License for the
-   specific language governing permissions and limitations under the License.
+   SPDX-License-Identifier: Apache-2.0
 */
-
 
 package com.example.handlingformsubmission;
 
@@ -21,13 +14,17 @@ import org.springframework.web.bind.annotation.PostMapping;
 
 @Controller
 public class GreetingController {
+    private final DynamoDBEnhanced dde;
+    private final PublishTextSMS msg;
 
     @Autowired
-    private DynamoDBEnhanced dde;
-
-    @Autowired
-    private PublishTextSMS msg;
-
+    GreetingController(
+        DynamoDBEnhanced dde,
+        PublishTextSMS msg
+    ) {
+        this.dde = dde;
+        this.msg = msg;
+    }
 
     @GetMapping("/")
     public String greetingForm(Model model) {
@@ -38,13 +35,11 @@ public class GreetingController {
     @PostMapping("/greeting")
     public String greetingSubmit(@ModelAttribute Greeting greeting) {
 
-        //Persist submitted data into a DynamoDB table using the Enhanced Client
+        // Persist submitted data into a DynamoDB table.
         dde.injectDynamoItem(greeting);
 
-        // Send a mobile notification
+        // Send a mobile notification.
         msg.sendMessage(greeting.getId());
-
         return "result";
     }
-
 }
