+++ javav2/usecases/creating_amazon_cognito_app/src/main/java/com/example/handlingformsubmission/GreetingItems.java
@@ -1,0 +1,55 @@
+/*
+   Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+   SPDX-License-Identifier: Apache-2.0
+*/
+
+package com.example.handlingformsubmission;
+
+import software.amazon.awssdk.enhanced.dynamodb.mapper.annotations.DynamoDbBean;
+import software.amazon.awssdk.enhanced.dynamodb.mapper.annotations.DynamoDbPartitionKey;
+
+@DynamoDbBean
+public class GreetingItems {
+
+    private String id;
+    private String name;
+    private String message;
+    private String title;
+
+    public GreetingItems()
+    {
+    }
+
+    public String getId() {
+        return this.id;
+    }
+
+    @DynamoDbPartitionKey
+    public void setId(String id) {
+        this.id = id;
+    }
+
+    public String getName() {
+        return this.name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public String getMessage(){
+        return this.message;
+    }
+
+    public void setMessage(String message){
+        this.message = message;
+    }
+
+    public String getTitle() {
+        return this.title;
+    }
+
+    public void setTitle(String title) {
+        this.title = title;
+    }
+}
