+++ javav2/usecases/creating_amazon_cognito_app/src/main/java/com/example/handlingformsubmission/PublishTextSMS.java
@@ -1,12 +1,6 @@
 /*
    Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
-   This file is licensed under the Apache License, Version 2.0 (the "License").
-   You may not use this file except in compliance with the License. A copy of
-   the License is located at
-    http://aws.amazon.com/apache2.0/
-   This file is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
-   CONDITIONS OF ANY KIND, either express or implied. See the License for the
-   specific language governing permissions and limitations under the License.
+   SPDX-License-Identifier: Apache-2.0
 */
 
 
@@ -31,7 +25,7 @@ public class PublishTextSMS {
                 .credentialsProvider(EnvironmentVariableCredentialsProvider.create())
                 .build();
         String message = "A new item with ID value "+ id +" was added to the DynamoDB table";
-        String phoneNumber="18195765654"; //Replace with a mobile phone number
+        String phoneNumber="Replace with a mobile phone number" ;
 
         try {
             PublishRequest request = PublishRequest.builder()
