+++ javav2/usecases/creating_amazon_cognito_app/src/main/java/com/example/handlingformsubmission/WebSecurityConfig.java
@@ -1,3 +1,8 @@
+/*
+   Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+   SPDX-License-Identifier: Apache-2.0
+*/
+
 package com.example.handlingformsubmission;
 
 import org.springframework.context.annotation.Configuration;
